-Wconversion for TAO folder, part 1

+++ src/TAO/Ledger/chainstate.cpp
@@ -54,7 +54,7 @@ namespace TAO
             if(hashBestChain.load() != hashLast.load())
             {
                 hashLast = hashBestChain.load();
-                nLastTime = runtime::unifiedtimestamp();
+                nLastTime = static_cast<uint32_t>(runtime::unifiedtimestamp());
             }
 
 
@@ -69,8 +69,8 @@ namespace TAO
         /* Flag to tell if initial blocks are downloading. */
         double ChainState::PercentSynchronized()
         {
-            uint32_t nChainAge = (runtime::unifiedtimestamp() - 20 * 60) - (config::fTestNet ? NEXUS_TESTNET_TIMELOCK : NEXUS_NETWORK_TIMELOCK);
-            uint32_t nSyncAge  = (stateBest.load().GetBlockTime() - (config::fTestNet ? NEXUS_TESTNET_TIMELOCK : NEXUS_NETWORK_TIMELOCK));
+            uint32_t nChainAge = (static_cast<uint32_t>(runtime::unifiedtimestamp()) - 20 * 60) - (config::fTestNet ? NEXUS_TESTNET_TIMELOCK : NEXUS_NETWORK_TIMELOCK);
+            uint32_t nSyncAge  = static_cast<uint32_t>(stateBest.load().GetBlockTime() - static_cast<uint64_t>(config::fTestNet ? NEXUS_TESTNET_TIMELOCK : NEXUS_NETWORK_TIMELOCK));
 
             return (100.0 * nSyncAge) / nChainAge;
         }
